// src/lib/types/supabase.types.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[];

export type Database = {
  public: {
    Tables: {
      albums: {
        Row: {
          artist: string;
          artist_id: string;
          cover_url: string | null;
          created_at: string;
          id: string;
          release_date: string | null;
          spotify_id: string;
          spotify_updated_at: string | null;
          title: string;
          tracks: Json;
          updated_at: string;
        };
        Insert: {
          artist: string;
          artist_id: string;
          cover_url?: string | null;
          created_at?: string;
          id?: string;
          release_date?: string | null;
          spotify_id: string;
          spotify_updated_at?: string | null;
          title: string;
          tracks?: Json;
          updated_at?: string;
        };
        Update: {
          artist?: string;
          artist_id?: string;
          cover_url?: string | null;
          created_at?: string;
          id?: string;
          release_date?: string | null;
          spotify_id?: string;
          spotify_updated_at?: string | null;
          title?: string;
          tracks?: Json;
          updated_at?: string;
        };
        Relationships: [];
      };
      follows: {
        Row: {
          created_at: string;
          follower_id: string;
          following_id: string;
        };
        Insert: {
          created_at?: string;
          follower_id: string;
          following_id: string;
        };
        Update: {
          created_at?: string;
          follower_id?: string;
          following_id?: string;
        };
        Relationships: [
          {
            foreignKeyName: "follows_follower_id_fkey";
            columns: ["follower_id"];
            isOneToOne: false;
            referencedRelation: "profiles";
            referencedColumns: ["id"];
          },
          {
            foreignKeyName: "follows_following_id_fkey";
            columns: ["following_id"];
            isOneToOne: false;
            referencedRelation: "profiles";
            referencedColumns: ["id"];
          }
        ];
      };
      profiles: {
        Row: {
          avatar_url: string | null;
          bio: string | null;
          created_at: string;
          full_name: string | null;
          id: string;
          updated_at: string;
          username: string;
        };
        Insert: {
          avatar_url?: string | null;
          bio?: string | null;
          created_at?: string;
          full_name?: string | null;
          id: string;
          updated_at?: string;
          username: string;
        };
        Update: {
          avatar_url?: string | null;
          bio?: string | null;
          created_at?: string;
          full_name?: string | null;
          id?: string;
          updated_at?: string;
          username?: string;
        };
        Relationships: [];
      };
      reviews: {
        Row: {
          album_id: string;
          created_at: string;
          favorite_track: string | null;
          id: string;
          rating: Database["public"]["Enums"]["rating_value"];
          review_text: string | null;
          updated_at: string;
          user_id: string;
        };
        Insert: {
          album_id: string;
          created_at?: string;
          favorite_track?: string | null;
          id?: string;
          rating: Database["public"]["Enums"]["rating_value"];
          review_text?: string | null;
          updated_at?: string;
          user_id: string;
        };
        Update: {
          album_id?: string;
          created_at?: string;
          favorite_track?: string | null;
          id?: string;
          rating?: Database["public"]["Enums"]["rating_value"];
          review_text?: string | null;
          updated_at?: string;
          user_id?: string;
        };
        Relationships: [
          {
            foreignKeyName: "reviews_album_id_fkey";
            columns: ["album_id"];
            isOneToOne: false;
            referencedRelation: "albums";
            referencedColumns: ["id"];
          },
          {
            foreignKeyName: "reviews_user_id_fkey";
            columns: ["user_id"];
            isOneToOne: false;
            referencedRelation: "profiles";
            referencedColumns: ["id"];
          }
        ];
      };
      top_albums: {
        Row: {
          album_id: string;
          created_at: string;
          id: string;
          position: number;
          updated_at: string;
          user_id: string;
        };
        Insert: {
          album_id: string;
          created_at?: string;
          id?: string;
          position: number;
          updated_at?: string;
          user_id: string;
        };
        Update: {
          album_id?: string;
          created_at?: string;
          id?: string;
          position?: number;
          updated_at?: string;
          user_id?: string;
        };
        Relationships: [
          {
            foreignKeyName: "top_albums_album_id_fkey";
            columns: ["album_id"];
            isOneToOne: false;
            referencedRelation: "albums";
            referencedColumns: ["id"];
          },
          {
            foreignKeyName: "top_albums_user_id_fkey";
            columns: ["user_id"];
            isOneToOne: false;
            referencedRelation: "profiles";
            referencedColumns: ["id"];
          }
        ];
      };
    };
    Views: {
      [_ in never]: never;
    };
    Functions: {
      [_ in never]: never;
    };
    Enums: {
      rating_value:
        | "0.5"
        | "1.0"
        | "1.5"
        | "2.0"
        | "2.5"
        | "3.0"
        | "3.5"
        | "4.0"
        | "4.5"
        | "5.0";
    };
    CompositeTypes: {
      [_ in never]: never;
    };
  };
};

type DefaultSchema = Database[Extract<keyof Database, "public">];
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof Database },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof Database;
  }
    ? keyof Database[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never
> = DefaultSchemaEnumNameOrOptions extends { schema: keyof Database }
  ? Database[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
  ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
  : never;